perf: バンドル最適化とパフォーマンス改善
- 重複するユーティリティ関数とエクスポートを完全に解消
- calculateMortgage関数の重複を削除（math.ts → mortgage-calculator.ts）
- generateRandomHex/generateUUID関数の重複を削除（password.ts → uuid.ts/random-number.ts）
- 時間フォーマット関数を統合（time-utils.tsに集約）
- 型定義の重複を解消：
  - ActivityLevel型の競合を修正（water-intake.ts: WaterActivityLevel）
  - WeightUnit型をwater-intake.tsに統一
  - Climate/SpecialCondition型の重複を削除
- テストファイルのimport pathを正しいソースファイルに修正
- ビルド時の重複インポート警告を完全に解消
- hexフォーマットを小文字に統一

バンドルサイズが削減され、コードの一貫性と保守性が向上

Closes #40
