+++ pages/tools/calorie-calculator.vue
@@ -195,10 +195,10 @@ import {
   formatGrams,
   getGoalDescription,
   type Goal,
-  type WeightUnit,
   type HeightUnit,
   type CalorieCalculatorResult,
 } from '~/utils/calorie-calculator'
+import type { WeightUnit } from '~/utils/water-intake'
 
 const gender = ref<Gender>('male')
 const age = ref<number>(30)
