+++ pages/tools/water-intake-calculator.vue
@@ -149,7 +149,7 @@ import {
   formatOunces,
   getHydrationTips,
   type WeightUnit,
-  type ActivityLevel,
+  type WaterActivityLevel,
   type Climate,
   type SpecialCondition,
   type WaterIntakeResult,
@@ -157,7 +157,7 @@ import {
 
 const weight = ref<number>(60)
 const weightUnit = ref<WeightUnit>('kg')
-const activityLevel = ref<ActivityLevel>('moderate')
+const activityLevel = ref<WaterActivityLevel>('moderate')
 const climate = ref<Climate>('temperate')
 const specialCondition = ref<SpecialCondition>('none')
 const exerciseMinutes = ref<number>(0)
