+++ utils/health.ts
@@ -11,24 +11,6 @@ export interface CalorieResult {
   description: string
 }
 
-export interface WaterIntakeResult {
-  baseWater: number
-  totalWater: number
-  totalWaterOz: number
-  activityAdjustment: number
-  climateAdjustment: number
-  exerciseAdjustment: number
-  specialConditionAdjustment: number
-  description: string
-  breakdown: {
-    base: string
-    activity: string
-    climate: string
-    exercise: string
-    specialCondition: string
-  }
-}
-
 export type Gender = 'male' | 'female'
 export type ActivityLevel =
   | 'sedentary'
@@ -36,8 +18,8 @@ export type ActivityLevel =
   | 'moderate'
   | 'active'
   | 'extra'
-export type Climate = 'temperate' | 'hot' | 'cold'
-export type SpecialCondition = 'none' | 'pregnancy' | 'breastfeeding'
+
+// Note: For water intake related types, use water-intake.ts
 
 export const calculateBMI = (weight: number, height: number): BMIResult => {
   if (weight <= 0 || height <= 0) {
