+++ utils/pomodoroTimer.ts
@@ -247,26 +247,8 @@ export function getPhaseDisplayName(phase: PomodoroPhase): string {
   }
 }
 
-export function formatTime(seconds: number): string {
-  const minutes = Math.floor(seconds / 60)
-  const remainingSeconds = seconds % 60
-  return `${minutes.toString().padStart(2, '0')}:${remainingSeconds.toString().padStart(2, '0')}`
-}
-
-export function formatDuration(minutes: number): string {
-  if (minutes < 60) {
-    return `${minutes}分`
-  }
-
-  const hours = Math.floor(minutes / 60)
-  const remainingMinutes = minutes % 60
-
-  if (remainingMinutes === 0) {
-    return `${hours}時間`
-  }
-
-  return `${hours}時間${remainingMinutes}分`
-}
+// Note: Time formatting functions are now available in time-utils.ts
+// Use formatPomodoroTime and formatPomodoroDuration from time-utils.ts
 
 export function getProgressPercentage(state: PomodoroState): number {
   if (state.totalTime === 0) return 0
