+++ utils/stopwatch.ts
@@ -12,12 +12,9 @@ export interface Lap {
   timestamp: Date
 }
 
-export interface FormattedTime {
-  hours: string
-  minutes: string
-  seconds: string
-  milliseconds: string
-}
+import { formatTimeToString } from './time-utils'
+
+// Note: Time utilities are now centralized in time-utils.ts
 
 export function createStopwatch(): StopwatchState {
   return {
@@ -94,57 +91,7 @@ export function getCurrentElapsedTime(state: StopwatchState): number {
   return Date.now() - state.startTime
 }
 
-export function formatTime(milliseconds: number): FormattedTime {
-  const totalMs = Math.abs(milliseconds)
-  const ms = totalMs % 1000
-  const totalSeconds = Math.floor(totalMs / 1000)
-  const seconds = totalSeconds % 60
-  const totalMinutes = Math.floor(totalSeconds / 60)
-  const minutes = totalMinutes % 60
-  const hours = Math.floor(totalMinutes / 60)
-
-  return {
-    hours: hours.toString().padStart(2, '0'),
-    minutes: minutes.toString().padStart(2, '0'),
-    seconds: seconds.toString().padStart(2, '0'),
-    milliseconds: Math.floor(ms / 10)
-      .toString()
-      .padStart(2, '0'),
-  }
-}
-
-export function formatTimeToString(milliseconds: number): string {
-  const formatted = formatTime(milliseconds)
-
-  if (parseInt(formatted.hours) > 0) {
-    return `${formatted.hours}:${formatted.minutes}:${formatted.seconds}.${formatted.milliseconds}`
-  }
-
-  return `${formatted.minutes}:${formatted.seconds}.${formatted.milliseconds}`
-}
-
-export function parseTimeString(timeString: string): number {
-  const parts = timeString.split(':')
-  let milliseconds = 0
-
-  if (parts.length === 3) {
-    // HH:MM:SS.MS format
-    const hours = parseInt(parts[0])
-    const minutes = parseInt(parts[1])
-    const [seconds, ms] = parts[2].split('.').map(s => parseInt(s))
-
-    milliseconds =
-      hours * 3600000 + minutes * 60000 + seconds * 1000 + (ms || 0) * 10
-  } else if (parts.length === 2) {
-    // MM:SS.MS format
-    const minutes = parseInt(parts[0])
-    const [seconds, ms] = parts[1].split('.').map(s => parseInt(s))
-
-    milliseconds = minutes * 60000 + seconds * 1000 + (ms || 0) * 10
-  }
-
-  return milliseconds
-}
+// Time formatting functions are now imported from time-utils.ts
 
 export function getBestLap(laps: Lap[]): Lap | null {
   if (laps.length === 0) {
